requirements.txt
spotipy>=2.11.0
protobuf==3.20.*
requests
python-dotenv
yt-dlp
bs4
imageio
PyNaCl
mutagen
typing-extensions
youtube-transcript-api
yt-dlp[default] --pre
google-generativeai
setuptools
pytz
pinecone[grpc]
deemix
pandas
httpx
discord
aiohttp

git+https://github.com/Shewiiii/librespot-python
git+https://github.com/Pycord-Development/pycord